feet: add error handler to login fetch
