+++ .env.example
@@ -5,3 +5,4 @@ FIREBASE_STORAGE_BUCKET=
 FIREBASE_MESSAGING_SENDER_ID=
 FIREBASE_APP_ID=
 FIREBASE_MEASUREMENT_ID=
+GOOGLE_APPLICATION_CREDENTIALS=service-account.json
