+++ .gitignore
@@ -22,3 +22,4 @@ logs
 .env
 .env.*
 !.env.example
+service-account.json
