+++ nuxt.config.ts
@@ -15,6 +15,9 @@ export default defineNuxtConfig({
     }
   ],
   vuefire: {
+    auth: {
+      enabled: true
+    },
     config: {
       apiKey: process.env.FIREBASE_API_KEY,
       authDomain: process.env.FIREBASE_AUTH_DOMAIN,
